Added Elixir Loqui client (#9)
* Added Elixir Loqui client

Added an elixir client and did some slight refactoring to the server
as well. The new client's parsing takes advantage of erlang's binary
optimizations to not split incoming binaries until all the packets are
parsed. This should have some impact on the server, so I changed it
there too.

Tested at 125,000 requests per second for a single client process with 100 workers against the go Loqui server.